Move qos.hpp header file to implementation folder
